chore(game): extract CombatManager from MainScene (#144)
Extract all combat and damage logic from MainScene into a dedicated
CombatManager class. Moves player attacks, super projectile handling,
enemy/boss attacks, bomb explosions, boss summon/ground pound, entity
death VFX, hit-stop, and chest interaction into the new manager.

MainScene reduced from ~2274 to ~1612 lines (-660 lines).
